chore: move ajax funcs to separate file
